hot fix

+++ 0x1A-hash_tables/4-hash_table_get.c
@@ -9,7 +9,7 @@ char *hash_table_get(const hash_table_t *ht, const char *key)
 		return (NULL);
 	}
 	index = key_index((const unsigned char *)key, ht->size);
-	if (index >= ht->size)
+	if ((long unsigned int)index >= ht->size)
 		return (NULL);
 
 	if ((ht->array)[index] == NULL)
